rpi4: Fix memory area overlay DTS
First fix is to change the memory area reserved for
the VC GPU from 64 -> 76 MB, as this is the default
value. Keeping that at 64 would have required everyone
to remember to set the "gpu_mem_XXXX" parameter
in config.txt.

Second fix is separating the disjunct memory areas
to separate nodes, as the generator Python scripts
in kernel/tools/ etc discarded the first area below 1GB
if all areas were defined in the same node.
The effect of different variations can be observed
in the project build folder:
build-folder/kernel/gen_headers/plat/machine/devices_gen.h

Third fix is adding the reserved memory area for the
GPU memory to avoid any possibility for collisions
between the kernel and VC.

Signed-off-by: Joonas Onatsu <joonas.onatsu@unikie.com>