kernel/tools: Fix handling of reserved memory
The set object of reserved regions,
given as a parameter to 'reserve_regions()'
function will become empty as the function
uses a 'pop()' operation on the object.

Later, 'reserved' is intended to be merged
with 'extra_reserved'. As the 'reserved' became
empty previously, the returned set contains
only 'extra_reserved' regions.

Fixed by handing the function a copy of the
original 'reserved' set.

Signed-off-by: Joonas Onatsu <joonasx@ssrc.tii.ae>